Made some file structure adjustments.
