+++ ParsonsFinalProject/ParsonsFinalProject.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27130.2036
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ParsonsFinalProject", "ConsoleApp1\ParsonsFinalProject.csproj", "{BBC79199-383B-4AA4-BB9C-4E90624CF416}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ParsonsFinalProject", "ParsonsFinalProject\ParsonsFinalProject.csproj", "{7F32CE20-123A-43E2-9D71-005790BDFEC1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{BBC79199-383B-4AA4-BB9C-4E90624CF416}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BBC79199-383B-4AA4-BB9C-4E90624CF416}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BBC79199-383B-4AA4-BB9C-4E90624CF416}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BBC79199-383B-4AA4-BB9C-4E90624CF416}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7F32CE20-123A-43E2-9D71-005790BDFEC1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7F32CE20-123A-43E2-9D71-005790BDFEC1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7F32CE20-123A-43E2-9D71-005790BDFEC1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7F32CE20-123A-43E2-9D71-005790BDFEC1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
