+++ ParsonsFinalProject/ParsonsFinalProject/ParsonsFinalProject.csproj
@@ -4,7 +4,7 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{B1A76183-2447-4093-9036-2CF5ADD76FF1}</ProjectGuid>
+    <ProjectGuid>{7F32CE20-123A-43E2-9D71-005790BDFEC1}</ProjectGuid>
     <OutputType>Exe</OutputType>
     <RootNamespace>ParsonsFinalProject</RootNamespace>
     <AssemblyName>ParsonsFinalProject</AssemblyName>
@@ -32,6 +32,9 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="NAudio, Version=1.8.4.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\NAudio.1.8.4\lib\net35\NAudio.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
@@ -41,15 +44,20 @@
     <Reference Include="System.Net.Http" />
     <Reference Include="System.Xml" />
     <Reference Include="VoiceRSS_TTS_SDK">
-      <HintPath>..\..\..\..\..\AppData\Local\Temp\voicerss_tts_dotnet-1\VoiceRSS_TTS_SDK.dll</HintPath>
+      <HintPath>.\VoiceRSS_TTS_SDK.dll</HintPath>
     </Reference>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="UserView.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
+    <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="ProjectIdeas.txt" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
